Tuning and "GAME OVER"

+++ inc/Variabel.h
@@ -9,7 +9,7 @@
 #define _VARIABEL_H_
 
 typedef struct {
-    int32_t x, y;
+    uint8_t x, y;
 } meteor_t;
 void gravityCheckBullet(bullet_t *b, meteor_t m);
 void gravityCheckShip(ship_t *ship, meteor_t m);
